Refactor prefect tests
